rag-poc-middleware-app/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
mongoengine>=0.28.2
pydantic>=2,<3
pydantic-settings>=2.3,<3
python-dotenv>=1.0.0
python-multipart==0.0.6
email-validator==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
pika==1.3.1
httpx==0.25.2
openai>=1.0.0
